agent/pyproject.toml
[project]
name = "luci-agent"
version = "0.1.0"
description = "Luci — Python Agent (orchestrator)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "pyjwt[crypto]>=2.10",
    "cryptography>=44",
    "psycopg[binary,pool]>=3.2",
    "redis>=5.2",
    "structlog>=24.4",
    "prometheus-client>=0.21",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "mypy>=1.13",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "openapi-python-client>=0.21",
    "httpx>=0.28",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
